+++ lib/clipboard.rb
@@ -1,15 +1,11 @@
 # frozen_string_literal: true
 
-require 'rbconfig'
-
 require_relative 'clipboard/version'
 require_relative 'clipboard/utils'
+require_relative 'clipboard/clipboard_load_error'
 
 module Clipboard
   extend self
-
-  class ClipboardLoadError < StandardError
-  end
 
   unless defined? Ocra # see gh#9
     autoload :Linux,          'clipboard/linux'
@@ -25,28 +21,7 @@ module Clipboard
 
   def self.implementation
     return @implementation if @implementation
-
-    os = case RbConfig::CONFIG['host_os']
-    when /mac|darwin/        then :Mac
-    when /linux|bsd/         then :Linux
-    when /mswin|mingw/       then :Windows
-    when /cygwin/            then :Cygwin
-    else
-      raise ClipboardLoadError, "Your OS(#{ RbConfig::CONFIG['host_os'] }) is not supported"
-    end
-
-    # Running additional check to detect if running in Microsoft WSL or Wayland
-    if os == :Linux
-      require "etc"
-      if Etc.respond_to?(:uname) && Etc.uname[:release] =~ /Microsoft/ # uname was added in ruby 2.2
-        os = :Wsl
-      # Only choose Wayland implementation if wl-copy is found, since xclip / xsel *might* work
-      elsif ENV["XDG_SESSION_TYPE"] == "wayland" && Utils.executable_installed?("wl-copy")
-        os = :LinuxWayland
-      end
-    end
-
-    @implementation = Clipboard.const_get(os)
+    @implementation = Clipboard.const_get(Utils.autodetect_implementation)
   rescue ClipboardLoadError => e
     $stderr.puts "#{e.message}\nUsing file-based (fake) clipboard" unless $VERBOSE == nil
     @implementation = Clipboard::File