+++ lib/clipboard/clipboard_load_error.rb
@@ -1,0 +1,4 @@
+module Clipboard
+  class ClipboardLoadError < StandardError
+  end
+end